rpc: Delete unnecessary server.rs.
